test(executor): split recovery hint and tool-loop stop reason tests into ctx-helpers file
